Comment fixes
