+++ assets/dev/scss/frontend/_div-block.scss
@@ -1,5 +1,5 @@
 // TODO remove in feature [ED-16336]
-.div-block {
+.e-div-block {
 	display: block;
 	padding-block-start: var(--padding-block-start);
 	padding-block-end: var(--padding-block-end);
