+++ modules/atomic-widgets/base/atomic-element-base.php
@@ -3,7 +3,6 @@
 namespace Elementor\Modules\AtomicWidgets\Base;
 
 use Elementor\Element_Base;
-use Elementor\Modules\AtomicWidgets\Controls\Section;
 use Elementor\Modules\AtomicWidgets\PropTypes\Concerns\Has_Atomic_Base;
 use Elementor\Modules\AtomicWidgets\Styles\Style_Schema;
 use Elementor\Modules\AtomicWidgets\Validators\Props_Validator;
@@ -27,44 +26,6 @@ abstract class Atomic_Element_Base extends Element_Base {
 
 		$this->version = $data['version'] ?? '0.0';
 		$this->styles = $data['styles'] ?? [];
-	}
-
-	private function get_valid_controls( array $schema, array $controls ): array {
-		$valid_controls = [];
-
-		foreach ( $controls as $control ) {
-			if ( $control instanceof Section ) {
-				$cloned_section = clone $control;
-
-				$cloned_section->set_items(
-					$this->get_valid_controls( $schema, $control->get_items() )
-				);
-
-				$valid_controls[] = $cloned_section;
-				continue;
-			}
-
-			if ( ! ( $control instanceof Atomic_Control_Base ) ) {
-				Utils::safe_throw( 'Control must be an instance of `Atomic_Control_Base`.' );
-				continue;
-			}
-
-			$prop_name = $control->get_bind();
-
-			if ( ! $prop_name ) {
-				Utils::safe_throw( 'Control is missing a bound prop from the schema.' );
-				continue;
-			}
-
-			if ( ! array_key_exists( $prop_name, $schema ) ) {
-				Utils::safe_throw( "Prop `{$prop_name}` is not defined in the schema of `{$this->get_name()}`." );
-				continue;
-			}
-
-			$valid_controls[] = $control;
-		}
-
-		return $valid_controls;
 	}
 
 	abstract protected function define_atomic_controls(): array;
