+++ modules/atomic-widgets/prop-types/concerns/has-atomic-base.php
@@ -2,7 +2,10 @@
 
 namespace Elementor\Modules\AtomicWidgets\PropTypes\Concerns;
 
+use Elementor\Modules\AtomicWidgets\Base\Atomic_Control_Base;
+use Elementor\Modules\AtomicWidgets\Controls\Section;
 use Elementor\Modules\AtomicWidgets\PropsResolver\Props_Resolver;
+use Elementor\Utils;
 
 if ( ! defined( 'ABSPATH' ) ) {
 	exit; // Exit if accessed directly.
@@ -11,6 +14,44 @@ if ( ! defined( 'ABSPATH' ) ) {
 trait Has_Atomic_Base {
 	public function has_widget_inner_wrapper(): bool {
 		return false;
+	}
+
+	private function get_valid_controls( array $schema, array $controls ): array {
+		$valid_controls = [];
+
+		foreach ( $controls as $control ) {
+			if ( $control instanceof Section ) {
+				$cloned_section = clone $control;
+
+				$cloned_section->set_items(
+					$this->get_valid_controls( $schema, $control->get_items() )
+				);
+
+				$valid_controls[] = $cloned_section;
+				continue;
+			}
+
+			if ( ! ( $control instanceof Atomic_Control_Base ) ) {
+				Utils::safe_throw( 'Control must be an instance of `Atomic_Control_Base`.' );
+				continue;
+			}
+
+			$prop_name = $control->get_bind();
+
+			if ( ! $prop_name ) {
+				Utils::safe_throw( 'Control is missing a bound prop from the schema.' );
+				continue;
+			}
+
+			if ( ! array_key_exists( $prop_name, $schema ) ) {
+				Utils::safe_throw( "Prop `{$prop_name}` is not defined in the schema of `{$this->get_name()}`." );
+				continue;
+			}
+
+			$valid_controls[] = $control;
+		}
+
+		return $valid_controls;
 	}
 
 	public function get_atomic_controls() {
